Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,7 +2,7 @@
 int main() {
     int a,b,c;
     scanf("%d %d %d",&a,&b,&c);
-    int sum=a+b+c;
+    float sum=a+b+c;
     float av=sum/3;
     printf("Average: %.2f",av);
     return 0;
